New Page Placeholders & Links
I added place holder pages for the attendance pages, the edit event page, and the page for editing the main public page. The links are also created to reach the respective pages from the buttons.

Note: still need to find a way to display the rsvp number on the event cards, and also need to make sure the edit event page pulls up the corresponding page for that specific event